Adapt talpid-windows-net to windows-sys 0.48
