+++ talpid-types/src/lib.rs
@@ -1,7 +1,5 @@
 #![deny(rust_2018_idioms)]
 
-use std::{error::Error, fmt, fmt::Write};
-
 #[cfg(target_os = "android")]
 pub mod android;
 pub mod net;
@@ -13,54 +11,5 @@ pub mod cgroup;
 #[cfg(target_os = "windows")]
 pub mod split_tunnel;
 
-/// Used to generate string representations of error chains.
-pub trait ErrorExt {
-    /// Creates a string representation of the entire error chain.
-    fn display_chain(&self) -> String;
-
-    /// Like [Self::display_chain] but with an extra message at the start of the chain
-    fn display_chain_with_msg(&self, msg: &str) -> String;
-}
-
-impl<E: Error> ErrorExt for E {
-    fn display_chain(&self) -> String {
-        let mut s = format!("Error: {self}");
-        let mut source = self.source();
-        while let Some(error) = source {
-            write!(&mut s, "\nCaused by: {error}").expect("formatting failed");
-            source = error.source();
-        }
-        s
-    }
-
-    fn display_chain_with_msg(&self, msg: &str) -> String {
-        let mut s = format!("Error: {msg}\nCaused by: {self}");
-        let mut source = self.source();
-        while let Some(error) = source {
-            write!(&mut s, "\nCaused by: {error}").expect("formatting failed");
-            source = error.source();
-        }
-        s
-    }
-}
-
-#[derive(Debug)]
-pub struct BoxedError(Box<dyn Error + 'static + Send>);
-
-impl fmt::Display for BoxedError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl Error for BoxedError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        self.0.source()
-    }
-}
-
-impl BoxedError {
-    pub fn new(error: impl Error + 'static + Send) -> Self {
-        BoxedError(Box::new(error))
-    }
-}
+mod error;
+pub use error::*;
